Include enum value in pure example (#58)
To demonstrate the code and the output.

This is documented in the derive crate but only the set up code and not
what the output looks like. I think it is useful to have an example of
the output too.

I'm happy to change anything about this.

Thank you again for the project!