Refactor StubInfoBuilder to reduce dependence on PyProject
To support generating stubs without a pyproject.toml file, it needs to be possible to build a StubInfo without one. The StubInfo only uses it to find the python root, so that value can be determined earlier in the process and passed along instead of the PyProject.

StubInfoBuilder is not exported, so its ::new() can be renamed without causing backwards incompatibility.