Enums with methods (#86)
Hello,

I was lacking the possibility of having methods and properties to enums,
so here is a proposal that adds this functionality. Additionally, the
documentation of documented members is now also present in the ".pyi"
output.

Here is a summary of the few changes:
- `MemberInfo` and `MemberDef` have a `doc` and a `is_property`
attribute.
- `StubInfoBuilder::add_methods` adds member properties for enums (but
not for classes, the behaviour remains the same)
- `StubInfoBuilder::add_methods` now looks for `struct_id` in
`module.enum_` in addition to `module.class`.

I would really like these changes to make it to a new release, please
let me know if additional changes are required to comply with the
current code base.