For rich enums, add Derive variant classes from the enum class.
This allows variants to be correctly returned from functions that
have a return type of the enum.

Avoids explicit setting of a base class in the macro by
creating an unqualified TypeInfo struct from the enum python class name
This avoids having to double-specify the enum python class name
when manually defining the stubs for a rich enum.