Derive macro for complex enums (#208)
Add new attribute macro `gen_stub_pyclass_complex_enum` that creates
internal classes for the variants of rich enums, that is, enums with
variants that have fields.

Also creates in type_info.rs PyComplexEnumInfo and VariantInfo for
manual creation of stub information for rich enums.

Both of these are handled by 'ClassDef', which can now have internal
classes and will write properly indented internal classes to the stub
file.

This PR also fixes the implementation on tuples so that tuples of length
one are supported.

---------

Co-authored-by: Erik von Reis <evonreis@caltech.edu>
Co-authored-by: James Gregory <james@jagregory.com>
Co-authored-by: Toshiki Teramura <toshiki.teramura@gmail.com>